Move CUDA stuff into CuStateVec
